本調檢查另外一个檔案
